@@ -2,23 +2,25 @@
   "author": "Kyle Krull",
   "description": "Remaps notes in MIDI tracks",
   "devDependencies": {
-    "@cucumber/cucumber": "^9.5.1",
-    "@cucumber/pretty-formatter": "^1.0.0",
-    "@tsconfig/node18": "^18.2.2",
+    "@cucumber/cucumber": "^11.1.1",
+    "@cucumber/pretty-formatter": "^1.0.1",
+    "@eslint/eslintrc": "^3.2.0",
+    "@eslint/js": "^9.17.0",
+    "@tsconfig/node18": "^18.2.4",
     "@types/chai": "^4.3.6",
-    "@types/node": "^20.8.2",
-    "@typescript-eslint/eslint-plugin": "^6.7.4",
-    "@typescript-eslint/parser": "^6.7.4",
+    "@types/node": "^22.10.2",
     "chai": "^4.3.10",
-    "eslint": "^8.50.0",
-    "husky": "^8.0.3",
-    "lint-staged": "^14.0.1",
-    "markdownlint-cli2": "^0.10.0",
-    "prettier": "3.0.3",
-    "prettier-plugin-organize-imports": "^3.2.3",
-    "ts-node": "^10.9.1",
+    "eslint": "^9.17.0",
+    "globals": "^15.14.0",
+    "husky": "^9.1.7",
+    "lint-staged": "^15.2.11",
+    "markdownlint-cli2": "^0.17.0",
+    "prettier": "3.4.2",
+    "prettier-plugin-organize-imports": "^4.1.0",
+    "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.2.2"
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.18.2"
   },
   "license": "MIT",
   "main": "index.js",
@@ -26,14 +28,14 @@
   "scripts": {
     "format": "prettier --ignore-unknown --write .",
     "format:check": "prettier --ignore-unknown --list-different .",
-    "lint": "eslint features features/**/*.cjs src",
+    "lint:check": "eslint features features/**/*.cjs src",
     "lint:fix": "eslint --fix features features/**/*.cjs src",
     "main:copy-chunks": "ts-node -r tsconfig-paths/register src/main/copy-chunks.ts",
     "main:copy-tracks": "ts-node -r tsconfig-paths/register src/main/copy-tracks.ts",
     "main:list-chunks": "ts-node -r tsconfig-paths/register src/main/list-chunks.ts",
     "main:list-events": "ts-node -r tsconfig-paths/register src/main/list-events.ts",
     "main:remap-events": "ts-node -r tsconfig-paths/register src/main/remap-events.ts",
-    "prepare": "husky install",
+    "prepare": "husky",
     "test": "cucumber-js -c features/cucumber.cjs",
     "test:ci": "cucumber-js -c features/cucumber.cjs --format html:output/test-cucumberjs.html",
     "test:focus": "cucumber-js -c features/cucumber.cjs --tags '@focus'",
