@@ -1,3 +1,7 @@
+import eslint from '@eslint/js';
+import globals from 'globals';
+import tsEslint from 'typescript-eslint';
+
 const allowUnusedIgnoreVariables = {
   'no-unused-vars': 'off', //avoid false positives from the base rule
   '@typescript-eslint/no-unused-vars': [
@@ -14,17 +18,16 @@ const youTryWritingStepDefinitionsIn80CharsOrLess = {
   'max-len': ['warn', { code: 100 }],
 };
 
-/* eslint-env node */
-module.exports = {
-  extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
-  env: {
-    node: true,
+export default tsEslint.config(
+  eslint.configs.recommended,
+  tsEslint.configs.recommended,
+  {
+    languageOptions: {
+      globals: { ...globals.node },
+    },
+    rules: {
+      ...allowUnusedIgnoreVariables,
+      ...youTryWritingStepDefinitionsIn80CharsOrLess,
+    },
   },
-  parser: '@typescript-eslint/parser',
-  plugins: ['@typescript-eslint'],
-  root: true,
-  rules: {
-    ...allowUnusedIgnoreVariables,
-    ...youTryWritingStepDefinitionsIn80CharsOrLess,
-  },
-};
+);
